+++ bill/delivery.go
@@ -23,6 +23,7 @@ const (
 	DeliveryTypeNote    cbc.Key = "note"
 	DeliveryTypeWaybill cbc.Key = "waybill"
 	DeliveryTypeReceipt cbc.Key = "receipt"
+	DeliveryTypeReturn  cbc.Key = "return"
 )
 
 // DeliveryTypes provides the list of supported delivery documents in GOBL.
@@ -75,6 +76,17 @@ var DeliveryTypes = []*cbc.Definition{
 			`),
 		},
 	},
+	{
+		Key: DeliveryTypeReturn,
+		Name: i18n.String{
+			i18n.EN: "Return Note",
+		},
+		Desc: i18n.String{
+			i18n.EN: here.Doc(`
+				A return note documents goods being sent back to the supplier.
+			`),
+		},
+	},
 }
 
 var isValidDeliveryType = cbc.InKeyDefs(DeliveryTypes)
@@ -87,7 +99,7 @@ type Delivery struct {
 	uuid.Identify
 
 	// Type of delivery document.
-	Type cbc.Key `json:"type" jsonschema:"title=Type" jsonschema_extras:"enum=advice,note,waybill,receipt"`
+	Type cbc.Key `json:"type" jsonschema:"title=Type" jsonschema_extras:"enum=advice,note,waybill,receipt,return"`
 	// Series is used to identify groups of deliveries by date, business area, project,
 	// type, customer, a combination of any, or other company specific data.
 	// If the output format does not support the series as a separate field, it will be